+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## (IN PROGRESS)
 
 * "Collapse all" accordions on "contact person" view not working. Refs UIORGS-231.
+* Contact persons email does not appear in the view. Refs UIORGS-232.
 
 ## [2.3.0](https://github.com/folio-org/ui-organizations/tree/v2.3.0) (2021-03-15)
 [Full Changelog](https://github.com/folio-org/ui-organizations/compare/v2.2.1...v2.3.0)